Board Detail UI to show/hide Update and Delete Board buttons based on Signed in User 'role' on board.

+++ src/components/BoardDetail/BoardDetail.jsx
@@ -2,24 +2,29 @@ import { useParams, useNavigate, Link } from 'react-router-dom'
 import { useState, useEffect } from 'react';
 import * as boardsAPI from '../../utilities/boards-api'
 
-export default function BoardDetail() {
+export default function BoardDetail({userProp}) {
     const navigate = useNavigate();
     const { boardName } = useParams()
     const boardNameActual = boardName ? boardName.replace(/-/g, ' ') : ''
 
-    const [userBoard, setUserBoard] = useState([])
+    const [userBoard, setUserBoard] = useState({})
 
     useEffect(function() {
         async function getBoard() {
-          const boards = await boardsAPI.getUserBoard(boardNameActual)
+            const boards = await boardsAPI.getUserBoard(boardNameActual)
           setUserBoard(boards)
         }
         getBoard()
     }, [boardNameActual])
 
+    
     async function handleDeleteClick() {
-        await boardsAPI.deleteUserBoard(boardNameActual)
-        navigate('/boards');
+        if (userProp._id === userBoard.author._id) {
+            await boardsAPI.deleteUserBoard(boardNameActual)
+            navigate('/boards');
+        } else {
+            alert("Only the author of a board can delete it.")
+        }
     }
 
     return (
@@ -45,10 +50,18 @@ export default function BoardDetail() {
                     )}
                 </div>
                 <p>{userBoard.createdAt}</p>
-                <Link to={`/boards/${boardName}/update`} >
-                    <button>Update Board</button>
-                </Link>
-                <button onClick={handleDeleteClick}>Delete Board</button>
+                
+                {userProp._id === userBoard.author?._id ? (
+                    <>
+                        <Link to={`/boards/${boardName}/update`} >
+                            <button>Update Board</button>
+                        </Link>
+                        <button onClick={handleDeleteClick}>Delete Board</button>
+                    </>
+                ) : (
+                <></>
+                )}
+
             </div>
         </>
     )
